added cast to integer to several hashtable removes unintentionally removing by index; added catch for concurrent modification exception in ar room draw; set portal activity to use material theme so buttons and text show up better on the surface view

+++ Source/Cypher/app/src/main/java/com/semaphore_soft/apps/cypher/game/Room.java
@@ -93,7 +93,7 @@ public class Room
     {
         if (residentActorIDs.contains(actorID))
         {
-            residentActorIDs.remove(actorID);
+            residentActorIDs.remove((Integer) actorID);
         }
     }
 
@@ -114,7 +114,7 @@ public class Room
     {
         if (residentEntityIDs.contains(entityID))
         {
-            residentEntityIDs.remove(entityID);
+            residentEntityIDs.remove((Integer) entityID);
         }
     }
